Refactor core middleware package into module
