--- misago/core/middleware/frontendcontext.py
@@ -1,7 +1,0 @@
-from django.utils.deprecation import MiddlewareMixin
-
-
-class FrontendContextMiddleware(MiddlewareMixin):
-    def process_request(self, request):
-        request.include_frontend_context = True
-        request.frontend_context = {}
