--- misago/core/middleware/threadstore.py
@@ -1,9 +1,0 @@
-from django.utils.deprecation import MiddlewareMixin
-
-from misago.core import threadstore
-
-
-class ThreadStoreMiddleware(MiddlewareMixin):
-    def process_response(self, request, response):
-        threadstore.clear()
-        return response
